+++ main_codemap/batterijen.py
@@ -8,7 +8,7 @@ class Batterijen:
         self.capaciteit = capaciteit
         self.resterende_capaciteit = capaciteit
         self.gelinkte_huizen = []
-        self.afstand_huizen = {}
+        self.afstand_huizen = []
         
         #to do: lijst van dicts maken met huis_id: huis object en afstand: afstande. 
     
@@ -17,9 +17,9 @@ class Batterijen:
         # itereer over ongekoppelde huizen, voeg deze toe aan dictionary en geef afstand tot batterij als waarde mee
         for huis in huizen:
             afstand = abs(huis.x_as - self.x_as) + abs(huis.y_as - self.y_as)
-            self.afstand_huizen[huis] = afstand
-            self.afstand_huizen = dict(sorted(self.afstand_huizen.items(), key=lambda x: x[1], reverse=True))
-            # print(f"afstand huizen: {self.afstand_huizen}")
+            self.afstand_huizen.append({'huis': huis, 'afstand': afstand})
+        self.afstand_huizen.sort(key=lambda x: x['afstand'], reverse=True)
+        print(self.afstand_huizen)
 
     def closest_house(self):
         distance = self.afstand_huizen
@@ -27,7 +27,3 @@ class Batterijen:
 
     def update_usage(self, output):
         self.resterende_capaciteit -= output
-
-    def clear_linked_houses(self):
-        new_dict = {key: value for key, value in self.afstand_huizen.items() if key.linked == False}
-        self.afstand_huizen = new_dict
