+++ main_codemap/smartgrid_algo_thom.py
