--- main_codemap/test.py
@@ -1,1 +1,0 @@
-"_" + 2
